Add component for handling unsaved changes
Sometimes users navigate away from a form before saving their changes. This can lead to users losing what they entered.

This JS warns the user if they attempt to leave the page without saving, and is based on the JS written for Apply for teacher training:
https://github.com/DFE-Digital/apply-for-teacher-training/blob/main/app/frontend/packs/warn-on-unsaved-changes.js

Add a component in the same pattern as Edge:

Usage in a prototype:
`<form data-module="warn-on-unsaved-changes">`